hook into reading a file to encode some new lines, move files around
